Redraw Dropdowns for QA
Added a line of code to allow the dropdown select of the QA posts which allows querying of monsters in said posts to be redrawn. This means that whenever a user queries a monster from the QA dropdown, it will immediately redraw and reset the selection to the empty default selection.

Minor additional commenting changes.